Enhance Wi-Fi and Bluetooth Management, Improve Service Initialization, and Fix Memory Handling
- Updated WifiManage to refresh the network list when scanning is not active and connection is possible.
- Added error handling in BluetoothHidHost to notify the UI of connection failures for non-HID devices.
- Improved DevelopmentService to ensure service is added if not found during lookup.
- Enhanced DisplayIdleService to ensure proper service initialization and loading.
- Refactored EspNow service to ensure it initializes correctly and handles peer additions.
- Updated MemoryCheckerService to adjust free memory thresholds for ESP32C5.
- Improved WebServerService to ensure proper initialization and settings loading.
- Enhanced Wifi service to manage user disconnections and auto-connect behavior more effectively.
- Added functionality to retrieve saved SSIDs from properties files in WifiApSettings.
- Improved WifiBootSplashInit to handle interrupted saves and ensure proper file recovery.
- Updated WifiSettings to support NVS for ESP_PLATFORM, ensuring settings are saved and loaded correctly.
- Refactored TouchCalibrationSettings to support NVS for ESP_PLATFORM, ensuring calibration settings are persisted.
- Fixed memory print stats to correctly format output for ESP targets.
- Enhanced properties file handling to recover from interrupted saves.
- Updated device.py to respect explicit board SDK configuration overrides for SPIRAM allocation.
- Bumped version to 0.8.1-ESP32-C5-dev.
- Added new provisioning file for cb2enterprises5G AP settings.

+++ TactilityKernel/source/memory.cpp
@@ -19,10 +19,12 @@ void memory_print_stats() {
 #ifdef ESP_PLATFORM
     size_t heap_free = heap_caps_get_free_size(MALLOC_CAP_INTERNAL);
     size_t heap_total = heap_caps_get_total_size(MALLOC_CAP_INTERNAL);
-    LOG_I(TAG, "Heap: %zu / %zu available", heap_free, heap_total);
+    // %zu is not supported by the reduced-size printf implementation used on ESP targets,
+    // so the values are cast and printed with %lu instead.
+    LOG_I(TAG, "Heap: %lu / %lu available", (unsigned long)heap_free, (unsigned long)heap_total);
     size_t ext_free = heap_caps_get_free_size(MALLOC_CAP_SPIRAM);
     size_t ext_total = heap_caps_get_total_size(MALLOC_CAP_SPIRAM);
-    LOG_I(TAG, "External: %zu / %zu available", ext_free, ext_total);
+    LOG_I(TAG, "External: %lu / %lu available", (unsigned long)ext_free, (unsigned long)ext_total);
 #endif
 }
